server/requirements.txt
fastapi
torch
torchvision
uvicorn
Pillow